+++ CHANGELOG.md
@@ -125,6 +125,7 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 - Updated `transfer_transaction_hbar.py` example to use `Hbar` objects instead of raw integers and added receipt checking with `ResponseCode` validation.([#1249](https://github.com/hiero-ledger/hiero-sdk-python/issues/1249))
 - Renamed `pr-missing-linked-issue.yml` and `pr_missing_linked_issue.js` to `bot-pr-missing-linked-issue.yml` and `bot-pr-missing-linked-issue.js` respectively. Enhanced LinkBot PR comment with clickable hyperlinks to documentation for linking issues and creating issues. (#1264)
 - Enhance assignment bot to guard against users with spam PRs `.github/scripts/bot-assignment-check.sh`
+- Enhance NodeAddress tests with additional coverage for proto conversion `tests/unit/node_address_test.py`
 
 ### Fixed
 - GFI workflow casing